refactor: move CeloConnectRedirect to pages app
