+++ src/providers/CeloProvider.tsx
@@ -1,9 +1,8 @@
 import { CeloProvider as ReactCeloProvider } from '@celo/react-celo';
 import '@celo/react-celo/lib/styles.css';
-import { useRouter } from 'next/router';
 import { PropsWithChildren } from 'react';
 import { networkConfig } from 'src/config/celo';
-import { AccountProvider, useAccountContext } from './AccountProvider';
+import { AccountProvider } from './AccountProvider';
 import { ExchangeProvider } from './ExchangeProvider';
 import { ThemeProvider } from './ThemeProvider';
 
@@ -26,29 +25,9 @@ export const CeloProvider = (props: PropsWithChildren) => {
     >
       <ThemeProvider>
         <AccountProvider>
-          <ExchangeProvider>
-            <CeloConnectRedirect>{props.children}</CeloConnectRedirect>
-          </ExchangeProvider>
+          <ExchangeProvider>{props.children}</ExchangeProvider>
         </AccountProvider>
       </ThemeProvider>
     </ReactCeloProvider>
   );
-};
-
-const CeloConnectRedirect = (props: PropsWithChildren) => {
-  const router = useRouter();
-  const { isConnected } = useAccountContext();
-
-  if (!isConnected && router.pathname !== '/connect') {
-    void router.push('/connect');
-
-    // Router is async. Show empty screen before redirect.
-    return null;
-  } else if (isConnected && router.pathname == '/connect') {
-    void router.push('/');
-
-    return null;
-  }
-
-  return <>{props.children}</>;
 };
